app/requirements.txt
numpy==1.22.4
opencv-contrib-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pytesseract==0.3.9
